move lib functions to their proper files
