+++ next.config.ts
@@ -22,6 +22,7 @@ const nextConfig: NextConfig = {
   },
   webpack: (config) => {
     config.externals = [...config.externals, { canvas: "canvas" }]; // required to make Konva & react-konva work
+    config.externals = [...config.externals, { sharp: "commonjs sharp" }];
     return config;
   },
 };
