+++ src/lib/session.ts
@@ -1,17 +1,10 @@
 import { cache } from "react";
 import { cookies } from "next/headers";
 
-import { sha256 } from "@oslojs/crypto/sha2";
-import {
-  encodeBase32LowerCaseNoPadding,
-  encodeHexLowerCase,
-} from "@oslojs/encoding";
-import type { Session } from "@prisma/client";
+import { encodeBase32LowerCaseNoPadding } from "@oslojs/encoding";
 
-import { prisma } from "@/lib/prisma";
+import { validateSessionToken } from "@/lib/data";
 import { SessionValidationResult } from "@/lib/types";
-
-const ONE_DAY_MS = 1000 * 60 * 60 * 24;
 
 export function generateSessionToken(): string {
   const bytes = new Uint8Array(20);
@@ -19,65 +12,6 @@ export function generateSessionToken(): string {
   const token = encodeBase32LowerCaseNoPadding(bytes);
 
   return token;
-}
-
-export async function createSession(
-  token: string,
-  userId: number,
-): Promise<Session> {
-  const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
-  const session: Session = {
-    id: sessionId,
-    userId,
-    expiresAt: new Date(Date.now() + ONE_DAY_MS * 30),
-  };
-  await prisma.session.create({
-    data: session,
-  });
-
-  return session;
-}
-
-export async function validateSessionToken(
-  token: string,
-): Promise<SessionValidationResult> {
-  const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
-  const result = await prisma.session.findUnique({
-    where: {
-      id: sessionId,
-    },
-    include: {
-      user: true,
-    },
-  });
-
-  if (result === null) {
-    return { session: null, user: null };
-  }
-  const { user, ...session } = result;
-
-  if (Date.now() >= session.expiresAt.getTime()) {
-    await prisma.session.delete({ where: { id: sessionId } });
-    return { session: null, user: null };
-  }
-
-  if (Date.now() >= session.expiresAt.getTime() - ONE_DAY_MS * 15) {
-    session.expiresAt = new Date(Date.now() + ONE_DAY_MS * 30);
-    await prisma.session.update({
-      where: {
-        id: session.id,
-      },
-      data: {
-        expiresAt: session.expiresAt,
-      },
-    });
-  }
-
-  return { session, user };
-}
-
-export async function invalidateSession(sessionId: string): Promise<void> {
-  await prisma.session.delete({ where: { id: sessionId } });
 }
 
 export async function setSessionTokenCookie(
